test(semantic tokens): split basic canary by construct area
The pre-refactor placeholder basic.cpp becomes two marker-free canaries
with clear scopes: declaration_tokens.cpp over declaration syntax and
expression_tokens.cpp over statements, expressions and attributes.
